requirements.txt
# Core dependencies
pynput>=1.7.6
sounddevice>=0.4.6
numpy>=1.21.0
openai-whisper>=20230314

# Additional dependencies for whisper
torch>=1.10.0
torchaudio>=0.10.0
numba>=0.56.4
more-itertools>=8.12.0
tiktoken>=0.3.1